line-bot/requirements.txt
flask
flask-cors
requests
line-bot-sdk>=3.0.0
psycopg2-binary
gunicorn
python-dotenv
